% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{data2}
\format{
A data frame with columns:
\describe{
\item{year}{year}
\item{dbh}{dbh}
\item{h}{h}
}
}
\source{
\strong{Gehrhardt, E. (1923) Ertragstafeln für Eiche, Buche, Tanne, Fichte und Riefer. Berlag von Julius Springer. Berlin. 46 pages. pp. 35-40.}
}
\usage{
data2
}
\description{
Gehrhardt extended Eichhorn's rule so that not only was the mean height directly related to the standing volume regardless of site class, but the mean height was directly related to the total volume production.

To work this out for yourself, look at the forester
}
\keyword{datasets}
